Update unit tests

+++ src/main/java/org/krishna/demo/service/impl/EmployeeServiceImpl.java
@@ -2,6 +2,7 @@ package org.krishna.demo.service.impl;
 
 import java.util.List;
 
+import org.apache.commons.lang.StringUtils;
 import org.krishna.demo.dao.EmployeeDao;
 import org.krishna.demo.exceptions.EmployeeNotFoundException;
 import org.krishna.demo.exceptions.EmployeeServiceException;
@@ -23,6 +24,9 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Override
 	public Employee getEmployee(Long id) {
+		if(null == id)
+			throw new EmployeeServiceException("Id cannot be null.");
+		
 		Employee emp = this.employeeDao.getEmployee(id);
 		if(null == emp)
 			throw new EmployeeNotFoundException();
@@ -31,6 +35,9 @@ public class EmployeeServiceImpl implements EmployeeService {
 
 	@Override
 	public List<Employee> findEmployees(String search) {
+		if(StringUtils.isEmpty(search))
+			throw new EmployeeServiceException("Search cannot be empty.");
+		
 		List<Employee> employees = this.employeeDao.findEmployees(search);
 		if(null == employees || employees.isEmpty())
 			throw new EmployeeNotFoundException();
@@ -40,17 +47,25 @@ public class EmployeeServiceImpl implements EmployeeService {
 	@Override
 	public Employee addEmployee(Employee employee) {
 		if(null == employee)
-			throw new EmployeeServiceException();
+			throw new EmployeeServiceException("Employee cannot be empty");
 		return this.employeeDao.addEmployee(employee);
 	}
 
 	@Override
 	public Employee updateEmployee(Long id, Employee employee) {
+		if(null == id)
+			throw new EmployeeServiceException("Id cannot be null.");
+		if(null == employee)
+			throw new EmployeeServiceException("Employee cannot be empty");
+		
 		return this.employeeDao.updateEmployee(id, employee);
 	}
 
 	@Override
 	public void deleteEmployee(Long id) {
+		if(null == id)
+			throw new EmployeeServiceException("Id cannot be null.");
+		
 		this.employeeDao.deleteEmployee(id);
 	}
